Updating version number script to handle README.
If the readme file contains the string "MPAS-v#.#" where # and # are
replaced with the major and minor version numbers respectively, the
version update script now updates the version numbers in the README.md
file as well as all Registry.xml files in the repository.

The script needs to be run from the same directory as the README.md file
in order to properly update all version numbers.