Add script to re-grid MPAS output to regular grid
This script takes a planar MPAS output file and creates a new file with
variables on a regular grid.  Functionality may be improved, but seems
to work fine based on limited testing.  It's not meant as a replacement
to visualizing native MPAS grids, but provides an option for using
traditional regular grid netCDF tools (like ncview).