Updating documentation generator.
Adding a function to break hyper links that are in long tables that
could be longer than the width of the column.

Also, var_structs are now referred to by a structname_struct.tex file,
rather than a structname.tex file. The old way conflicted if the
struct name was the same as a namelist record name.