Fixing some issues:
Fixing the complete cell check in mpas_mesh_converter.
Also, making a cell be complete only if it has 90% of 2*Pi radians worth
of edges.

Adding debugging information for periodicity cheking.

Making periodicity range be 90% of the reference, as opposed to 50%.

Fixing an issue with cullCell field being filled with 1 when missing as
opposed to 0.