Adding a script to define cullMask for LI grids
Currently has 3 options for how to define the mask.  User can edit the
script to select to pick them.  Script works but doesn't have a lot of
bells and whistles.
This is basically copied with some modifications from the old script
define_keepCellMask.py, which will be deleted in a future commit.