Fixing an issue with verticesOnCell
Previously some vertices were duplicated in the verticesOnCell list.
This was due to edges that had a missing vertex. Missing vertices should
not be added to the verticesOnCell list, but also if an edge is missing
a vertex, the cell is going to be marked as "incomplete" for removal.

Now, vertices are only added to verticesOnCell if they are not missing.