Make CISM mesh unstructured for ESMF
The grid areas for the CISM grid had lots of values close to 0 and the
interpolation had weird artifacts in it.  It must be that my
specification of a structured SCRIP grid is still inadequate.

However, switching to making the CISM grid appear to be unstructured
eliminates the problems and makes not obvious impact on performance, so
I'll just use that instead...