viz block script: fix random colors, add stats
This fixes the recent implementation of random colors for the blocks in
the visualize_blocks.py script.  The previous implementation hardcoded
the number of random colors to be 256, which made some blocks the same
color when there was more than 256 blocks.

This commit also adds min/mean/max number of cells per block info to
stdout, and it adds a check to make sure nCells equals the number of
lines in the block file.