triangle_to_netcdf efficiency update
Do all calculations in memory and then write out to file at end.  This
is much faster for moderate to large datasets than doing all the
calculations directly to disk.  This could lead to memory problems on
huge meshes, and there may be some things that can be done to calculate
only some variables at a time and then delete them from memory.