periodic_general: add data density function option
There is an option in Params.txt to enable data density function.

If chosen, the program will look for a file called density.nc that has
dimensions: x, y and variables: x, y, density.  This is a regular grid
file so density variable has dimensions (x,y).

Right now it applies the density as uniform within each grid cell, but
it would be easier to add a bilinear interpolation function.  If the
point to be processed is outside the range of density grid, the nearest
value is applied.