Add barycentric interpolation to script to add CISM IC to MPASLI
Now there are three methods of interpolation:
b=bilinear, d=barycentric, e=ESMF

The barycentric version takes longer on init because it has to set up
interpolation weights.  But the actual interpolation is two orders of
magnitude faster than bilinear.  Whether that translates to an overall
speedup depends on the size of the meshes and the number of fields to
interpolate.