create LI grid script: add more stdout, speedup large grids
This commit adds additional print statements to
create_landice_grid_from_generic_MPAS_grid.py that are useful for
communicating what is happening when making large grids.

Included in this is a 'counter' showing the mesh fields as they are
copied.

Also, each mesh field is synced and deleted after it is created which
should improve performance for large meshes.