Adjust density value for MC point generation
The power of 0.5 is needed to be applied to the density value
when generating MC points to get the proper density of points.
The logic is that grid spacing in 2d is approximately equal
to density to the 1/4 power.  However, cell area is
approximately grid spacing squared, leading to a 1/2 power.