Add submit script for completing a LI mesh
This script takes a generic MPAS mesh and does all the steps that would
make a culled landice grid with initial conditions that is ready to be
used for simulations.

See new file mpas_mesh_to_landice_ic_batch.sh for details.  It is
primarily intended for use when extra memory is needed and a cluster has
to be used.  It is basically a template and will need to be modified for
any specific usage.