Add .m script for preparing pattyn temp field for interpolation
This commit adds a Matlab script used for preparing the 3d AIS
temperature field from Frank Pattyn for interpolation onto an MPAS mesh.

It includes:
* renaming/organizing fields to follow CISM conventions since the
interpolation script currently assumes that format
* Adding a layer of smoothed data in no data (non-ice) areas to avoid
the issue of interpolating junk into cells that don't match the ice
extent of this dataset.
To be added:
* Filtering to remove 'hot spot' artifacts present in this dataset.

Though this script is kind of a one-off specific to this dataset, I'm
committing here for reference should these steps be useful for
preprocessing other datasets in the future.