Update distance calc for points, and add point mask
This commit updates the distance calculation for computing cells that
contain points to use the angle of the dot product, rather than the dot
product.

Additionally, this commit adds the ability to write a field that is
cellPointMask which flags cells that contain points (to easily visualize
which cells contain points).