Add transfer of properties to masks file
This commit adds the ability for the properties of a feature to be
transferred into the NetCDF file as fields.

An example would be if a region has an attribute named depth, this would
result in the creation of a depthRegions field in the netCDF file, that
would contain the value of this property.