Add option to write data to combined file
With the combine flag, data are written to files fieldsOnCells.pvd,
fieldsOnEdges.pvd and fieldsOnVerties.pvd, with each file in the
time series containing all fields (whether time dependent or not)
in a vtp file.  This allows for combining and filtering fields
within Paraview (e.g. with the Compute filter).

Removes global_time_indices, which was simply range(nTimes) and
therefore does not need to be stored.

When files are not combined, time-independent files are no longer
written out to pdv files (undoing this change from the previous
commit).