Add script for extracting Paraview topography data
This merge adds a script that extracts geometry for plotting
MPAS fields with varying depth (presumably topography) as well
as fields sampled using a specified index field on cells.

Currently, only fields on cells are supported.  In paraview, fields
on cells will have constant values but will be interpolated over
rectangular steps between cells at edges.

Usage is similar to paraview_vtk_field_extractor.py.  New flags are:
-t, --topo_dim: specifies the vertical dimension of the model (e.g.
   nVertLevels in MPAS-Ocean)
-i, --topo_cell_index: specifies an index array on cells that corresponds
   to the location of the topography (e.g. maxLevelCell in MPAS-Ocean).
   If this flag is excluded, each cell is sampled in the bottom level of
   topo_dim.

results are stored in topoFieldsOnCells or timeDependentTopoFieldsOnCells
and staticTopoFieldsOnCells (depending on the --combine flag), so as
not to conflict with fieldsOnCells, etc. that may be extracted separatedly.