Add a script for extracting Paraview transect geometry
This merge adds a script, paraview_vtk_transect_extractor.py,
that can be used to extract data along transects specified by
a NetCDF file produced using the MaskCreator tool.

The syntax is similar to paraview_vtk_field_extractor.py, except
that only fields on cells are currently supported and the following
new flags have been added:
-r, --transects_file: a NetCDF file produce by the MaskCreator that
  contains one or more transects
-t, --transect_dim: the depth dimension for transects (e.g. nVertLevels
  in MPAS-Ocean)
-n, --min_level_cell: a cell index array specifying the first valid
  index in each column (0 by default)
-x, --max_level_cell: a cell index array specifiying the last valid
  index in each column (transect_dim-1 by default), e.g. maxLevelCell
  in MPAS-Ocean
-l, --layer_thickness: the name of the field where layer thicknesses
  are stored, used to compute the cumSumLayerThickness field that
  can be used in Paraview to compute the vertical coordinate of each
  transect.

The script averages layerThickness from cell centers to edges (or used
the layerThickness of the valid cell for edges and layers where only
neighboring cell is valid, e.g. at stair steps in the topography).
Then, a cumulative sum of layer thickness is computed and stored in
cumSumLayerThickness for use in Paraview.