Fix computing zInterface (was cumSumLayerThickness)
This merge adds flags for two new (optional) fields that can be
used to specify either z_min (the elevation/depth of the lower
interface of the first layer) or z_max (the elevation/depth of the
upper interface of the last layer).  These and also layer_thickness
can be specified with a minus sign as needed.  For example, in
MPAS-O, one could either specify:
--layer_thickness=-layerThickness --z_min=ssh
or
--layer_thickness=-layerThickness --z_max=-bottomDepth
both of which should produce the same result.  In MPAS-O, layer_thickness
is negative so that increasing layer index corresponds to decreasing
zInterface, whereas the opposite convention is (probably) the case
in various other MPAS cores.

The output variable for transect layer interfaces is now zInterface.