Add a utility library to paraview extractor
The utility library has several subroutines that will also be
used by scripts for extracting topography and transects in paraview.

A check has been added to make sure to use only the first
occurance of a time if times are not unique across files.

Add a flag for an output directory name.