Fix distance2 calculation in mpas_draw; fixes periodicity viz on narrow meshes
distance2 is used to determine which connectivity should be ignored
because it goes around the periodicity.  The calculation was incorrectly
using the square of the difference between points.  In meshes that
have a large aspect ratio, this leads to connectivity being included
that should not be, and that renders the mesh very incorrectly.

This commit changes the pow() to be the absolute value.  Visualization
of narrow periodic meshes now works properly.