Add support for index arrays on extra dimensions
This merge adds the ability to extract fields on a given dimension
using an index array.  This is convenient in MPAS-O, where fields
at the ocean bottom are at index maxLevelCell.

Example usage with the -d flag:

... -d nVertLeves=0,maxLevelCell

This will extract any fields that include dimension nVertLevels at
both the top index and at vertical index maxLevelCell (the
bottom of the ocean).