Generalize interp script to support MPASLI format as input
This commit modifies
	modified:   interpolate_cism_grid_to_mpas_grid.py
so that in can use either a CISM format file or an MPASLI format file as
the input file that provides the source data for interpolation.

For MPAS format files, currently only built-in barycentric interpolation
is supported.  ESMF regridding may also be possible, but I have not
looked into it.

The code has been kept general as much as possible to support
interpolation from both types of files in one script.  However the
script name (with "cism" in it) and command line options (short-form)
have been retained without modification to allow
backwards compatibility and to avoid having to adjust the test case
infrastructure.