interpolate script: remove try/except, just check for var names
The try/except around each variable mostly just hid errors.  This change
replaces it with a check for the input and output file variable names
and gracefully skips that variable if either is missing.  If the fields
do exist, there are no longer try/except around the interpolation
itself, so errors will present themselves.