Fixed interpolation for roundoff-level differences in vertical coordinates
I fixed a problem that can arise when interpolating 3D fields in the vertical
between CISM and MPAS.  Suppose the MPAS grid has a minimum layer at 0.0500000000,
and the input CISM file has a minimum layer at 0.0500000001.
Previously, the code punted because the minimum CISM layer was greater than the
minimum MPAS layer (and similarly for max layer values).  Now, the script
checks to see whether the values are very close (within 1.e-6).  If so,
it slightly adjusts the input CISM layer so that vertical interpolation
can proceed.