LI define_cullMask: ignore invalid cells
This commit adds in a check to ignore invalid cells.  Without this, the
-1 value for invalid cells gets interpreted as a valid index by python
and may or may not refer to a cell with ice.  If it is a cell with ice,
then all the cells along the edge of the mesh appear to have neighbors
with ice and get kept in the mask.