Correct history attribute in some grid_gen scripts
This commit fixes the history attribute as applied to the
landice grid tools.  NCO adds newest history to the beginning, and it
includes the date/time the command was executed.  I've made the
scripts in planar_grid_transformations conform to that convention.

I've also deleted appending the history attribute in
create_SCRIP_files/create_SCRIP_file_from_MPAS_mesh.py because the
source MPAS file is not actually modified by the script.