Add lon-lat mode to vtk extractor
With the --lonlat flag, fields are extracted in lon-lat space
(in degrees) instead of cartesian coordinates that plot on the
sphere.

If --lonlat is specified, only those cells (or edges or vertices)
that cross +/-180 will have extra vertices added to the vtk
geometry.

Blocking has also been removed from geometry extraction calculations
because this greatly simplifies the code.  If this proves to be
too memory intensive for large meshes, this should be reevaluated
later.

This merge also introduces a helper function for fixing lon-lat
vertices, allowing better code reuse.