Import evtk instead of pyevtk, if available
Only if the evtk package is not found do we use the old pyevtk
package.  The evtk package is being built from evtk source as
part of newer e3sm-unified packages and is on the e3sm channel.
pyevtk does not seem to reliably support python 3 and some
versions seem to have bugs in the setup script.