Fixes grid_rotate overwrite bug
When the same name was given for the source and destination file, grid_rotate
would overwrite the source file with the rotation, instead of failing.

This fix allows grid_rotate to exit when the source filename is the same as the
destination filename as well as any other cp error.