Add error handling for sphereRadius=0
Add simple error handling to avoid division by zero when sphereRadius field is set to 0,
which is commonly the case for planar meshes. Setting to 1 in this case ensures that the
grid_area values in the output SCRIP file take on the same values as the areaCell field
in the input MPAS input mesh file.