Remove shallow-water IC fields from periodic hex meshes
The periodic_hex mesh generator previously added fields to the grid.nc
file to serve as placeholders for ICs for the shallow-water model.
However, these fields are not part of the MPAS mesh specification, and
have been removed. Specifically, the following fields are deleted in
this commit:

  fEdge
  fVertex
  h_s
  u
  uBC
  v
  h
  vh
  circulation
  vorticity
  ke
  tracers

Additionally, the dimensions nVertLevels and nTracers have also been
removed.