Restructure for better use as python module
By restructuring the file, the function `make_periodic_planar_hex_mesh`
can be called from another python script to generate a mesh object
for further manipulation.  Eventually, this should help with
making a non-periodic version of the mesh, perhaps with another
tool.  It would also allow for culling cells or other operations
without the need to write out and read back in the mesh or to
call this script as a subprocess of another script.