Fixes output of x_period and y_period
For periodic grids, x and y_period where being read in as character
strings and then converted to doubles for output in the mesh.  The read
in as a string caused only the first two decimals to be read of y_period
and x_period, which breaks periodicity in MPAS.

This commit reads in x_period and y_period as doubles and skips the
conversions.