Make separate mpich, openmpi and "nompi" compass
The mpich version does not work on cori or grizzly compute nodes,
but does work on anvil/blues compute nodes.  While it works on
all login nodes, we don't want to run mpi jobs on those nodes.

Also, constrain pyremap in anticipation of major, incompatible
changes in v0.1.0 of that package.