Crop to a region in the ParaView extractor
If a geojson file (or FeatureCollection) is provided to the
ParaView VTK extractor, use conversion tools to create a mask,
cull the mesh to that mask, and extract the requested fields in
the masked region.

The default behavior has been changed so that `all`, `allOnCells`,
`allOnVertices`, and `allOnEdges` refer only to variables in the
time-series files by default, and will include variables from
the mesh file only if `--include_mesh_vars` is specified.

Some shared functionality has been pulled out into helper
functions.