Add function for converting MPAS meshes to viz triangles
Each pair of adjacent vertices on an MPAS cell is joined to the
cell center to create a triangle.  Interpolation weights and cell
indices are computed so that MPAS fields at cell centers can either
be plotted as "flat" to show the individual cell polygons or
"gouraud" to smoothly interpolate between values from the closest
cell centers.