exo to nc conversion: fix beta extrapolation
The old method for extrapolating beta used the grounded ice extent as
the source region for extrapolation.  However, Albany optimization
includes one extended cell beyond that, which was getting clobbered by
the extrapolation step.  This commit fixes that by defining the source
region as the cells that have beta>0 in the exodus file (i.e., keep any
locations where the optimization returned a beta value and extrapolate
out from there.)

This commit also disables the smoothing of the beta extrapolation, which
seems to provide hit or miss results.