Fix grid_imask in scrip_from_mpas
Previously, the code for computing grid_imask was erroneously
within an if statement that did not get called for cells with
the maximum number of vertices.

The code for filling in vertices has also been rearranged to loop
over vertices on a cell (typically 7) rather than cells (typically
tens of thousands to millions) for efficiency and to set a better
example for future codes that may be build off of this one.