MALI interp: allow generic filetypes to work with ESMF interp method
This allows this script to be used to interpolate from other E3SM
components (e.g. ELM) to the MALI grid using the ESMF weight files used
in the E3SM coupler.  In such a situation, the source file is neither
MPAS nor CISM format, but the ESMF method does not need any information
about the file format and will still work.  This commit relaxes some
logic to allow this workflow.