Fix maksing for signed distance functions
Previously, we were buffering shapes by a full grid cell,
but this was only necessary because we had the wrong origin for
rasterio's rasterize function.  If the origin is offset by
half a pixel to account for the fact that rasterio pixel corners
are half a pixel offset from the lon/lat that we want (at pixel
centers), we only need to buffer shapes by a tiny amount (say
1e-10 degrees) to ensure that lon/lat points on the domain
boundary are included within shapes that also touch the domain
boundaries.