Add check for required time dimension if a range of times is included
The interp script can optionally take a range of time levels to
interpolate.  Currently, if the expected time dimension is not in the
source file, the script will just interpolate the initial time level
from the source into all time levels of the destination.  This commit
fixes this behavior by aborting with an error in that situation.