Add thickness mask to exo thickness conversion
When the exo conversion script converts thickness, care has to be taken
to not add/change thickness to the extended row of cells that exists at
the calving front.  This commit avoids this by masking out the values
copied over from the Exodus file.  I also added a warning statement,
because we have not really tested this workflow out yet.  Once we are
routinely using the thickness adjustment capability, this message should
be removed.