Fix vertical layer value used for surfaceTemperature
The existing code was assuming stride was the number of vertical layers.
This change removes that assumption by using the number of vertical
layers as determined from the dimensions of the temperature field.

I also added an explicit declaration of the vertical layer used for
basalTemperature.  This doesn't change the value, but hopefully make the
code more readable.