Generalize angleEdge for planar meshes in MpasMeshConverter
In MpasMeshConverter, the planar mesh calculation for angleEdge assumes
the connectivity of periodic_hex such that angleEdge would always be
between 0 and 180.

This commit eliminates that assumption and makes the angleEdge
calculation applicable to any mesh.  This is necessary for angleEdge to
be correct for planar meshes made with JIGSAW, for example.