In grid_rotate/Makefile, set the Fortran compiler using 'nc-config --fc'
Previously, the name of the Fortran compiler was hard-wired to 'gfortran'. Even
in cases where grid_rotate was being built with the GNU compilers, this could
fail with undefined references, for example, if the NetCDF library was build
with parallel I/O support (in which case, we need to link with the MPI wrapper).

Now, the value of the FC variable in the top-level Makefile is set to the output
of 'nc-config --fc'.

One potential issue worth noting is that, while the Fortran compiler can change
depending on the output of nc-config, the Fortran compiler flags (FFLAGS) are
still hard-wired in the top-level Makefile.