Add argument to write_netcdf for engine
Performance for the default 'netcdf4' engine has found to be poor
for large files, and 'scipy' seems to perform better.  For now,
'netcdf4' remains the default but the option to change is now
provided.

To aid in this process, default values have been defined in the
mpas_tools.io module that can be modified once, rather than
each time `write_netcdf` is called.