Add an iceMask field to the interpolation script
Add an iceMask field to the interpolation script, which can be used
for trimming down fields to the original extent of ice after
interpolating. This is useful when fields must be extrapolated before
using the interpolation script in order to avoid including zeros, NaNs,
or large missing values in the interpolation at the ice edge.