Add analysis script for plotting regional stats
This script is modified from the global stats script.  It plots three
different figures, each with subpanels for each region.  The first is
overall mass change statistics.  The second is the grounded mass
balance, and the third is the floating mass balance.  It is possible to
plot up to four runs, each with a different linestyle, but because of
the number of lines on each plot, plotting multiple runs is not pretty.