Make all processing steps optional and modular
This required removing the dependence of the flux processing on the
state processing output.  The only dependence was on the mask, so I
added a mask calculation to the flux module.  I also modified that mask
calculation to generate the maximum ice extent over the entire averaging
period instead of just using a snapshot at the end of the averaging
period.