Correct time averaging of flux variables, simplify code
This commit corrects the time averaging of flux variables, which previously
was not including the Jan 1 entry of the following calendar year.  That entry
technically belongs to the previous year for flux calculation.  This commit does
this by revamping the way time is handled in the flux time averaging.  This also
corrects the calculation of time bounds.

This commit also:
* stops calling the translate_GL_and_calving_flux_edge2cell routine, which we no
longer want to use.  The GL flux is now calculated correctly in MALI.  The calving
flux calculation should also be moved to MALI, but has not been implemented yet, because
we are not running with calving yet.  Eliminating the call to this function also allows
us to eliminate some ncks and shutil steps, because they are no longer needed.  That
also speeds things up, because some of those were kind of slow.  The code here is updated
to use the new MALI GL flux variable on cells.
* Enables dH/dt calculation.