Update time avg to use a separate output dataset
Before the time averager was writing the output to the input dataset
and then saving it as a new file.  After my recent modification of the
time averaging, this caused a problem, because the input dataset has
more time levels (non-time-averaged) than the output (time-averaged),
which led to a bunch of empty time levels in the processed files.

Now I create an xarray dataset from scratch, only including the
time-averaged fields, plus a few other neeed variables.  This solves the
problem, and also makes the code cleaner.