Workaround xarray data type problems
xarray assumes any variable with units of days since something are a
timedelta type, which is problematic.  The workaround I previously used
failed around 2280, apparently due to overflow.  I think timedelta type
uses units of nanoseconds or something and therefore can only capture a
relatively short length of time in years, which turns out to be just a
bit too short for our needs.  To get around this, I used a workaround
that deletes the units of daysSinceStart before it's read in.  However
that workaround broke xtime, so there are further changes to read and
use xtime correctly.