scalar fields: handle flux and state time levs separately, hist different
For processing the scalar fields, the state and flux fields need to have
the number of output time levels handled differently from each other.
Also, the hist experiment needs to be handlded differently than the
projection experiments.  This commit updates the time level management
of the scalar files to achieve the correct time handling.