mpas atmosphere mesh tool update.
This utility produces both a "scrip" (depricated) file and an esmf
mesh file compatible with CESM2.

Corrected the computation of the refinement factor and also have the
computation use the mesh density function.  The refinement factor is
used by the CESM/CAM topo tool to determine how to average the
topography and compute the subgrid topography variables for the CAM6
gravity-wave drag scheme.