Move mask creation to before interpolation in a separate output file
This commit creates 2 masks about the extent of ocean data and moves
them to before interpolation so that MALI has access to the interpolated
mask values, rather than trying to construct them after the fact.  The
masks get interpolated in a separate file from the TF/melt data.

It is not yet clear if there will be a single mask definition that is
appropriate for all applications, so this approach forces the MALI user
to make a conscious decision about how to define the mask.

The two masks are:
* the entire ocean domain
* the open ocean