Add script for adding mask to scrip file for ocn-glc fjord coupling
This script defines a mask of ocean cells that are deeper than a
threshold. This is to be used for updating a scrip file so that it only
remaps from ocean cells deeper than that threshold. The resulting
mapping file can be used as an OCN2GLC_TF_SMAPNAME mapping file in E3SM.
Note that config_2d_thermal_forcing_depth in the MPAS-Ocean namelist
needs to match the depth used here!