Add 3d ocean mask generation
This is a mask on the MALI mesh with the ISMIP6-ocean vertical
coordinate that indicates where valid MPAS-Ocean data exists (including
inside cavities).

To support this, a notable change is that nan is no longer inserted in
the TF field outside of the valid range.  Instead data is vertically
extrapolated and one should rely on orig3dOceanMask to identify where
the TF data can be used.