Simplify time-averaging function
Now that the year loop has been moved to the main driver, the time
averaging function can be simplified to only handle the single case of
time-averaging one year.  The year loop had to be moved to the main
driver to avoid running out of memory when processing many years.

Note there still is code in the function for the case of not
time-averaging.  The non-time-averaged case cannot currently be handled
by the script, and this is left as a placeholder until later when the
driver is restructured to support non-time-averaging.