Support larger mesh sizes in the NetCDF-C mesh converter
The `nCells`, `nEdges` and `nVertices` dimensions now are of
type `size_t` in the c++ code.  Products of these dimensions with
other dimensions (`maxEdges2`, `two`, `vertexDegree`, etc.) are
computed as `size_t` variables and then used to allocate arrays
to avoid `int` overflow.

These changes are likely not sufficient to allow meshes with more
than ~2 billion edges (or vertices or cells) because index arrays
are still stored as `int` type.