Remove encoding unlimited_dims if Time not present
When xarray 2025.08.0 opens a file with an empty unlimited
Time dimension but no variables with Time, it drops `Time` as a
dimension but keeps it in `unlimited_dims`.  This is probably
not an expected behavior but it's easy enough to address by
setting `unlimited_dims = None` if `Time` is not present.