Ensure `Time` dim is added to landice initial condition
MPAS meshes include an empty, unlimited `Time` dimension but the
latest NetCDF4 drops this dimension since no variable use it.
the `create_from_generic_mpas_grid()` funciton in `landice.create`
was taking advantage of `Time` being a dimension in the MPAS mesh
but this should no longer be assumed.  This merge fixes that
assumption by explicitly adding `Time` to the list of dimensions
to create if it is not already in the input dataset.